Add optional project selector to Add Story form
The project ID field is now optional. If it is left blank a new dropdown
will appear in the Add Story form allowing a project to be selected.

This change depends on an unreleased change in sentry master, so bump the
dependency to >5.4.5 (the current release).

See: https://github.com/getsentry/sentry/pull/864